FEAT: #11 POST user comment API

+++ src/main/java/com/unicon/rebit/yongginae/userComment/UserCommentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface UserCommentRepository extends JpaRepository<UserComment, Integer> {
+public interface UserCommentRepository extends JpaRepository<UserComment, Long> {
     // user_id로 userComment 값 가져오기
     List<UserComment> findByUser_id(Long id);
